+++ packages/vm/tests/api/utils.ts
@@ -17,9 +17,9 @@ export function setupVM(opts: VMOpts & { genesisBlock?: Block } = {}) {
     opts.blockchain = new Blockchain({
       db,
       validateBlocks: false,
-      validatePow: false,
+      validateConsensus: false,
       common,
-      genesisBlock,
+      genesisBlock
     })
   }
   return new VM(opts)